+++ lib/screens/seller/canteen_seller.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:hello_world/home_page.dart';
+import 'package:hello_world/screens/buyer/home_page.dart';
 
 class CanteenSeller extends StatefulWidget {
   const CanteenSeller({Key? key}) : super(key: key);
