+++ lib/screens/seller/mess_seller.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:hello_world/home_page.dart';
+import 'package:hello_world/screens/buyer/home_page.dart';
 
 class MessSeller extends StatefulWidget {
   const MessSeller({Key? key}) : super(key: key);
